Dealt with sudden removal of alg field on itsme JWK
This is a historic commit, it is not neccessary as node-jose correctly
deals with missing alg fields.

Original body:
node-jose seems to initialize its alg fields with an empty string and
consequently fails to match a search for an algorithm to a key without
alg. I'll investigate the cause and possible remedies and create an
issue in node-jose if so required.